move to domain utils
