--- server/src/utils.ts
@@ -1,3 +1,0 @@
-export const handlePromiseError = <T>(promise: Promise<T>): void => {
-  promise.catch((error) => console.error(`[utils.ts]:handlePromiseError ${error}`, error));
-};
